pyproject.toml
[tool.poetry]
name = "jquants-derivatives"
version = "0.1.0-alpha"
description = ""
authors = ["driller"]
readme = "README.md"
packages = [{include = "jquants_derivatives"}]

[tool.poetry.dependencies]
python = "^3.10"
jquants-api-client = "^1.2.0"
plotly = "^5.14.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.23.1"


[tool.poetry.group.test.dependencies]
pytest = "^7.3.1"
python-dotenv = "^1.0.0"
mypy = "^1.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
